Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpParameters-headerExtensions-expected.txt


FAIL setParameters() with modified headerExtensions should reject with InvalidModificationError assert_equals: Expect dictionary.rtcp to be plain object expected "object" but got "undefined"
